Add support for all missing command line arguments accepted by Kiota
The following parameters were added:
 - typeAccessModifier
 - backingStore
 - excludeBackwardCompatible
 - additionalData
 - structuredMimeTypes
 - disableValidationRules
 - disableSslValidation

Also, the types for the following parameters were made more precise:
 - cleanOutput: String -> boolean
 - clearCache: String -> boolean
 - kiotaLogLevel: String -> enum

Finally, extraArguments was added, which allows any argument to be
passed directly to Kiota. This will help if new parameters are added in
future versions of Kiota and the plugin hasn't been updated yet.